Add SplashPresenter
business logics are moved to presenter class.
