+++ CoffeeShopFinder/CoffeeShopFinder/UI/Base/BasePresenter.swift
@@ -1,0 +1,13 @@
+//
+//  BasePresenter.swift
+//  CoffeeShopFinder
+//
+//  Created by Joey Lee on 14/4/18.
+//  Copyright © 2018 Joey Lee. All rights reserved.
+//
+
+import UIKit
+
+class BasePresenter: NSObject {
+
+}
